requirements.txt
Flask
flask-cors
python-dotenv
pymongo
dnspython
gunicorn
